Refatoração na organização e listagem de comandos
