+++ src/commands/utility/commands-functions.utility.ts
@@ -1,39 +1,38 @@
 import { downloadMediaMessage, WASocket } from "baileys"
-import { Bot } from "../interfaces/bot.interface.js"
-import { Group } from "../interfaces/group.interface.js"
-import { Message } from "../interfaces/message.interface.js"
-import * as Whatsapp from '../lib/whatsapp.js'
-import { buildText, messageErrorCommandUsage} from "../lib/util.js"
+import { Bot } from "../../interfaces/bot.interface.js"
+import { Group } from "../../interfaces/group.interface.js"
+import { Message } from "../../interfaces/message.interface.js"
+import * as Whatsapp from '../../lib/whatsapp.lib.js'
+import { buildText, messageErrorCommandUsage} from "../../lib/util.lib.js"
 import { audioLibrary, generalLibrary, imageLibrary } from "@victorsouzaleal/biblioteca-lbot"
-import getCommands from "./list.commands.js"
-
+import { commandsUtility } from "./commands-list.utility.js"
 
 export async function animesCommand(client: WASocket, botInfo: Bot, message: Message, group? : Group){
-    const commandsData = getCommands(botInfo)
+    const utilityCommands = commandsUtility(botInfo)
     const animes = await generalLibrary.animeReleases()
-    let replyText = commandsData.utility.animes.msgs.reply_title
+    let replyText = utilityCommands.animes.msgs.reply_title
 
     animes.forEach((anime)=>{
-        replyText += buildText(commandsData.utility.animes.msgs.reply_item, anime.name.trim(), anime.episode, anime.url)
+        replyText += buildText(utilityCommands.animes.msgs.reply_item, anime.name.trim(), anime.episode, anime.url)
     })
 
     await Whatsapp.replyText(client, message.chat_id, replyText, message.wa_message, {expiration: message.expiration})
 }
 
 export async function mangasCommand(client: WASocket, botInfo: Bot, message: Message, group? : Group){
-    const commandsData = getCommands(botInfo)
+    const utilityCommands = commandsUtility(botInfo)
     const mangas = await generalLibrary.mangaReleases()
-    let replyText = commandsData.utility.mangas.msgs.reply_title
+    let replyText = utilityCommands.mangas.msgs.reply_title
 
     mangas.forEach((manga)=>{
-        replyText += buildText(commandsData.utility.mangas.msgs.reply_item, manga.name.trim(), manga.chapter, manga.url)
+        replyText += buildText(utilityCommands.mangas.msgs.reply_item, manga.name.trim(), manga.chapter, manga.url)
     })
 
     await Whatsapp.replyText(client, message.chat_id, replyText, message.wa_message, {expiration: message.expiration})
 }
 
 export async function brasileiraoCommand(client: WASocket, botInfo: Bot, message: Message, group? : Group){
-    const commandsData = getCommands(botInfo)
+    const utilityCommands = commandsUtility(botInfo)
     let seriesSupported = ['A', 'B']
     let serieSelected : "A" | "B"
 
@@ -51,13 +50,13 @@ export async function brasileiraoCommand(client: WASocket, botInfo: Bot, message
 
     const [round] = rounds.filter(round => round.rodada_atual === true)
     const {partidas: matches} = round
-    let replyText = buildText(commandsData.utility.brasileirao.msgs.reply_title, serieSelected)
+    let replyText = buildText(utilityCommands.brasileirao.msgs.reply_title, serieSelected)
 
-    replyText += commandsData.utility.brasileirao.msgs.reply_table_title
+    replyText += utilityCommands.brasileirao.msgs.reply_table_title
 
     table.forEach(team =>{
         replyText += buildText(
-            commandsData.utility.brasileirao.msgs.reply_table_item,
+            utilityCommands.brasileirao.msgs.reply_table_item,
             team.posicao,
             team.nome,
             team.pontos,
@@ -66,11 +65,11 @@ export async function brasileiraoCommand(client: WASocket, botInfo: Bot, message
         )
     })
 
-    replyText += "\n" + commandsData.utility.brasileirao.msgs.reply_round_title
+    replyText += "\n" + utilityCommands.brasileirao.msgs.reply_round_title
 
     matches.forEach(match =>{
         replyText += buildText(
-            commandsData.utility.brasileirao.msgs.reply_match_item,
+            utilityCommands.brasileirao.msgs.reply_match_item,
             match.time_casa,
             match.time_fora,
             match.data,
@@ -83,16 +82,16 @@ export async function brasileiraoCommand(client: WASocket, botInfo: Bot, message
 }
 
 export async function encurtarCommand(client: WASocket, botInfo: Bot, message: Message, group? : Group){
-    const commandsData = getCommands(botInfo)
+    const utilityCommands = commandsUtility(botInfo)
 
     if (!message.args.length) throw new Error(messageErrorCommandUsage(botInfo, message))
 
     const url = await generalLibrary.shortenUrl(message.text_command)
-    await Whatsapp.replyText(client, message.chat_id, buildText(commandsData.utility.encurtar.msgs.reply, url), message.wa_message, {expiration: message.expiration})
+    await Whatsapp.replyText(client, message.chat_id, buildText(utilityCommands.encurtar.msgs.reply, url), message.wa_message, {expiration: message.expiration})
 }
 
 export async function upimgCommand(client: WASocket, botInfo: Bot, message: Message, group? : Group){
-    const commandsData = getCommands(botInfo)
+    const utilityCommands = commandsUtility(botInfo)
     
     if (message.quotedMessage?.type !== 'imageMessage' && message.type !== 'imageMessage') throw new Error(messageErrorCommandUsage(botInfo, message))
    
@@ -104,23 +103,23 @@ export async function upimgCommand(client: WASocket, botInfo: Bot, message: Mess
     }
 
     let imageUrl = await imageLibrary.uploadImage(imageBuffer)
-    await Whatsapp.replyText(client, message.chat_id, buildText(commandsData.utility.upimg.msgs.reply, imageUrl), message.wa_message, {expiration: message.expiration})
+    await Whatsapp.replyText(client, message.chat_id, buildText(utilityCommands.upimg.msgs.reply, imageUrl), message.wa_message, {expiration: message.expiration})
 }
 
 export async function filmesCommand(client: WASocket, botInfo: Bot, message: Message, group? : Group){
-    const commandsData = getCommands(botInfo)
+    const utilityCommands = commandsUtility(botInfo)
     let movieTrendings = await generalLibrary.moviedbTrendings("movie")
-    await Whatsapp.replyText(client, message.chat_id, buildText(commandsData.utility.filmes.msgs.reply, movieTrendings), message.wa_message, {expiration: message.expiration})
+    await Whatsapp.replyText(client, message.chat_id, buildText(utilityCommands.filmes.msgs.reply, movieTrendings), message.wa_message, {expiration: message.expiration})
 }
 
 export async function seriesCommand(client: WASocket, botInfo: Bot, message: Message, group? : Group){
-    const commandsData = getCommands(botInfo)
+    const utilityCommands = commandsUtility(botInfo)
     let movieTrendings = await generalLibrary.moviedbTrendings("tv")
-    await Whatsapp.replyText(client, message.chat_id, buildText(commandsData.utility.series.msgs.reply, movieTrendings), message.wa_message, {expiration: message.expiration})
+    await Whatsapp.replyText(client, message.chat_id, buildText(utilityCommands.series.msgs.reply, movieTrendings), message.wa_message, {expiration: message.expiration})
 }
 
 export async function rbgCommand(client: WASocket, botInfo: Bot, message: Message, group? : Group){
-    const commandsData = getCommands(botInfo)
+    const utilityCommands = commandsUtility(botInfo)
 
     if (!message.isMedia && !message.isQuoted) throw new Error(messageErrorCommandUsage(botInfo, message))
 
@@ -129,42 +128,42 @@ export async function rbgCommand(client: WASocket, botInfo: Bot, message: Messag
         wa_message: (message.isQuoted)? message.quotedMessage?.wa_message : message.wa_message
     }
 
-    if (!messageData.type || !messageData.wa_message) throw new Error(commandsData.utility.rbg.msgs.error_message)
+    if (!messageData.type || !messageData.wa_message) throw new Error(utilityCommands.rbg.msgs.error_message)
 
-    if (messageData.type != "imageMessage") throw new Error(commandsData.utility.rbg.msgs.error_only_image)
+    if (messageData.type != "imageMessage") throw new Error(utilityCommands.rbg.msgs.error_only_image)
 
-    await Whatsapp.replyText(client, message.chat_id, commandsData.utility.rbg.msgs.wait, message.wa_message, {expiration: message.expiration})
+    await Whatsapp.replyText(client, message.chat_id, utilityCommands.rbg.msgs.wait, message.wa_message, {expiration: message.expiration})
     let imageBuffer = await downloadMediaMessage(messageData.wa_message, "buffer", {})
     let replyImageBuffer = await imageLibrary.removeBackground(imageBuffer)
     await Whatsapp.replyFileFromBuffer(client, message.chat_id, 'imageMessage', replyImageBuffer, '', message.wa_message, {expiration: message.expiration})
 }
 
 export async function tabelaCommand(client: WASocket, botInfo: Bot, message: Message, group? : Group){
-    const commandsData = getCommands(botInfo)
+    const utilityCommands = commandsUtility(botInfo)
     const replyText = await generalLibrary.symbolsASCI()
-    await Whatsapp.replyText(client, message.chat_id, buildText(commandsData.utility.tabela.msgs.reply, replyText), message.wa_message, {expiration: message.expiration})
+    await Whatsapp.replyText(client, message.chat_id, buildText(utilityCommands.tabela.msgs.reply, replyText), message.wa_message, {expiration: message.expiration})
 }
 
 export async function letraCommand(client: WASocket, botInfo: Bot, message: Message, group? : Group){
-    const commandsData = getCommands(botInfo)
+    const utilityCommands = commandsUtility(botInfo)
 
     if (!message.args.length) throw new Error(messageErrorCommandUsage(botInfo, message))
 
     const musicLyrics = await generalLibrary.musicLyrics(message.text_command)
-    const replyText = buildText(commandsData.utility.letra.msgs.reply, musicLyrics.title, musicLyrics.artist, musicLyrics.lyrics)
+    const replyText = buildText(utilityCommands.letra.msgs.reply, musicLyrics.title, musicLyrics.artist, musicLyrics.lyrics)
     await Whatsapp.replyFile(client, message.chat_id, 'imageMessage', musicLyrics.image, replyText, message.wa_message, {expiration: message.expiration})
 }
 
 export async function ouvirCommand(client: WASocket, botInfo: Bot, message: Message, group? : Group){
-    const commandsData = getCommands(botInfo)
+    const utilityCommands = commandsUtility(botInfo)
 
-    if (!process.env.dg_secret_key?.trim()) throw new Error(commandsData.utility.ouvir.msgs.error_key)
+    if (!process.env.dg_secret_key?.trim()) throw new Error(utilityCommands.ouvir.msgs.error_key)
     if (!message.isQuoted || message.quotedMessage?.type != 'audioMessage') throw new Error(messageErrorCommandUsage(botInfo, message))
-    if (message.quotedMessage?.media?.seconds && message.quotedMessage?.media?.seconds > 90) throw new Error(commandsData.utility.ouvir.msgs.error_audio_limit)
+    if (message.quotedMessage?.media?.seconds && message.quotedMessage?.media?.seconds > 90) throw new Error(utilityCommands.ouvir.msgs.error_audio_limit)
 
     let audioBuffer = await downloadMediaMessage(message.quotedMessage.wa_message, "buffer", {})
     let replyText = await audioLibrary.audioTranscription(audioBuffer, {deepgram_secret_key : process.env.dg_secret_key?.trim()})
-    await Whatsapp.replyText(client, message.chat_id, buildText(commandsData.utility.ouvir.msgs.reply, replyText), message.quotedMessage.wa_message, {expiration: message.expiration})
+    await Whatsapp.replyText(client, message.chat_id, buildText(utilityCommands.ouvir.msgs.reply, replyText), message.quotedMessage.wa_message, {expiration: message.expiration})
 }
 
 export async function audioCommand(client: WASocket, botInfo: Bot, message: Message, group? : Group){
@@ -179,7 +178,7 @@ export async function audioCommand(client: WASocket, botInfo: Bot, message: Mess
 }
 
 export async function traduzCommand(client: WASocket, botInfo: Bot, message: Message, group? : Group){
-    const commandsData = getCommands(botInfo)
+    const utilityCommands = commandsUtility(botInfo)
     const languageSupported = ["pt", "es", "en", "ja", "it", "ru", "ko"]
     let languageTranslation : string
     let textTranslation : string[] | string
@@ -197,17 +196,17 @@ export async function traduzCommand(client: WASocket, botInfo: Bot, message: Mes
     }
 
     if (!languageSupported.includes(languageTranslation)){
-        await Whatsapp.replyText(client, message.chat_id, commandsData.utility.traduz.msgs.error, message.wa_message, {expiration: message.expiration})
+        await Whatsapp.replyText(client, message.chat_id, utilityCommands.traduz.msgs.error, message.wa_message, {expiration: message.expiration})
         return
     }
 
     const replyTranslation = await generalLibrary.translationGoogle(textTranslation as string, languageTranslation as "pt" | "es" | "en" | "ja" | "it" | "ru" | "ko")
-    const replyText = buildText(commandsData.utility.traduz.msgs.reply, textTranslation as string, replyTranslation)
+    const replyText = buildText(utilityCommands.traduz.msgs.reply, textTranslation as string, replyTranslation)
     await Whatsapp.replyText(client, message.chat_id, replyText, message.wa_message, {expiration: message.expiration})
 }
 
 export async function vozCommand(client: WASocket, botInfo: Bot, message: Message, group? : Group){
-    const commandsData = getCommands(botInfo)
+    const utilityCommands = commandsUtility(botInfo)
     const languageSupported = ["pt", 'en', 'ja', 'es', 'it', 'ru', 'ko', 'sv']
     let languageVoice: string
     let textVoice : string[] | string
@@ -222,54 +221,54 @@ export async function vozCommand(client: WASocket, botInfo: Bot, message: Messag
         textVoice = textVoice.join(" ")
     }
 
-    if (!languageSupported.includes(languageVoice)) throw new Error(commandsData.utility.voz.msgs.error_not_supported)
-    if (!textVoice) throw new Error(commandsData.utility.voz.msgs.error_text)
-    if (textVoice.length > 500) throw new Error(commandsData.utility.voz.msgs.error_text_long)
+    if (!languageSupported.includes(languageVoice)) throw new Error(utilityCommands.voz.msgs.error_not_supported)
+    if (!textVoice) throw new Error(utilityCommands.voz.msgs.error_text)
+    if (textVoice.length > 500) throw new Error(utilityCommands.voz.msgs.error_text_long)
 
     const replyAudioBuffer = await audioLibrary.textToVoice(languageVoice as "pt" | "es" | "en" | "ja" | "it" | "ru" | "ko" | "sv", textVoice)
     await Whatsapp.replyFileFromBuffer(client, message.chat_id, 'audioMessage', replyAudioBuffer, '', message.wa_message, {expiration: message.expiration, mimetype: 'audio/mpeg'})
 }
 
 export async function noticiasCommand(client: WASocket, botInfo: Bot, message: Message, group? : Group){
-    const commandsData = getCommands(botInfo)
+    const utilityCommands = commandsUtility(botInfo)
 
     const newsList = await generalLibrary.newsGoogle()
-    let replyText = commandsData.utility.noticias.msgs.reply_title
+    let replyText = utilityCommands.noticias.msgs.reply_title
 
     for(let news of newsList){
-        replyText += buildText(commandsData.utility.noticias.msgs.reply_item, news.title, news.author, news.published, news.url)
+        replyText += buildText(utilityCommands.noticias.msgs.reply_item, news.title, news.author, news.published, news.url)
     }
 
     await Whatsapp.replyText(client, message.chat_id, replyText, message.wa_message, {expiration: message.expiration})
 }
 
 export async function calcCommand(client: WASocket, botInfo: Bot, message: Message, group? : Group){
-    const commandsData = getCommands(botInfo)
+    const utilityCommands = commandsUtility(botInfo)
 
     if (!message.args.length) throw new Error(messageErrorCommandUsage(botInfo, message))
     
     const calcResult = await generalLibrary.calcExpression(message.text_command)
-    const replyText = buildText(commandsData.utility.calc.msgs.reply, calcResult)
+    const replyText = buildText(utilityCommands.calc.msgs.reply, calcResult)
     await Whatsapp.replyText(client, message.chat_id, replyText, message.wa_message, {expiration: message.expiration})
 }
 
 export async function pesquisaCommand(client: WASocket, botInfo: Bot, message: Message, group? : Group){
-    const commandsData = getCommands(botInfo)
+    const utilityCommands = commandsUtility(botInfo)
 
     if (!message.args.length) throw new Error(messageErrorCommandUsage(botInfo, message))
 
     let webSearchList = await generalLibrary.webSearchGoogle(message.text_command)
-    let replyText = buildText(commandsData.utility.pesquisa.msgs.reply_title, message.text_command)
+    let replyText = buildText(utilityCommands.pesquisa.msgs.reply_title, message.text_command)
 
     for(let search of webSearchList){
-        replyText += buildText(commandsData.utility.pesquisa.msgs.reply_item, search.title, search.url)
+        replyText += buildText(utilityCommands.pesquisa.msgs.reply_item, search.title, search.url)
     }
 
     await Whatsapp.replyText(client, message.chat_id, replyText, message.wa_message, {expiration: message.expiration})
 }
 
 export async function moedaCommand(client: WASocket, botInfo: Bot, message: Message, group? : Group){
-    const commandsData = getCommands(botInfo)
+    const utilityCommands = commandsUtility(botInfo)
     const supportedCurrencies = ["dolar" , "iene" , "euro" , "real"]
 
     if (message.args.length != 2) throw new Error(messageErrorCommandUsage(botInfo, message))
@@ -279,23 +278,23 @@ export async function moedaCommand(client: WASocket, botInfo: Bot, message: Mess
     if (!supportedCurrencies.includes(currencySelected)) throw new Error(messageErrorCommandUsage(botInfo, message))
 
     let convertData = await generalLibrary.convertCurrency(currencySelected as "dolar" | "euro" | "real" | "iene", parseInt(valueSelected))
-    let replyText = buildText(commandsData.utility.moeda.msgs.reply_title, convertData.currency, convertData.value)
+    let replyText = buildText(utilityCommands.moeda.msgs.reply_title, convertData.currency, convertData.value)
 
     for(let convert of  convertData.convertion){
-        replyText += buildText(commandsData.utility.moeda.msgs.reply_item, convert.convertion_name, convert.value_converted_formatted, convert.currency, convert.updated)
+        replyText += buildText(utilityCommands.moeda.msgs.reply_item, convert.convertion_name, convert.value_converted_formatted, convert.currency, convert.updated)
     } 
 
     await Whatsapp.replyText(client, message.chat_id, replyText, message.wa_message, {expiration: message.expiration})
 }
 
 export async function climaCommand(client: WASocket, botInfo: Bot, message: Message, group? : Group){
-    const commandsData = getCommands(botInfo)
+    const utilityCommands = commandsUtility(botInfo)
 
     if (!message.args.length) throw new Error(messageErrorCommandUsage(botInfo, message))
 
     let wheatherResult = await generalLibrary.wheatherInfo(message.text_command)
 
-    let replyText = buildText(commandsData.utility.clima.msgs.reply,
+    let replyText = buildText(utilityCommands.clima.msgs.reply,
         message.text_command,
         wheatherResult.location.name,
         wheatherResult.location.region,
@@ -310,7 +309,7 @@ export async function climaCommand(client: WASocket, botInfo: Bot, message: Mess
     )
 
     wheatherResult.forecast.forEach((forecast)=>{
-        replyText += buildText(commandsData.utility.clima.msgs.reply_forecast,
+        replyText += buildText(utilityCommands.clima.msgs.reply_forecast,
             forecast.day,
             forecast.max,
             forecast.min,
@@ -326,12 +325,12 @@ export async function climaCommand(client: WASocket, botInfo: Bot, message: Mess
 }
 
 export async function dddCommand(client: WASocket, botInfo: Bot, message: Message, group? : Group){
-    const commandsData = getCommands(botInfo)
+    const utilityCommands = commandsUtility(botInfo)
     let dddSelected : string | undefined
 
     if (message.isQuoted){
         let internationalCode = message.quotedMessage?.sender.slice(0,2)
-        if (internationalCode != "55") throw new Error(commandsData.utility.ddd.msgs.error)
+        if (internationalCode != "55") throw new Error(utilityCommands.ddd.msgs.error)
         dddSelected = message.quotedMessage?.sender.slice(2,4)
     } else if (message.args.length){
         dddSelected = message.text_command
@@ -340,17 +339,17 @@ export async function dddCommand(client: WASocket, botInfo: Bot, message: Messag
     if (!dddSelected) throw new Error(messageErrorCommandUsage(botInfo, message))
 
     let dddResult = await generalLibrary.infoDDD(dddSelected)
-    const replyText = buildText(commandsData.utility.ddd.msgs.reply, dddResult.state, dddResult.region)
+    const replyText = buildText(utilityCommands.ddd.msgs.reply, dddResult.state, dddResult.region)
     await Whatsapp.replyText(client, message.chat_id, replyText, message.wa_message, {expiration: message.expiration})
 }
 
 export async function qualmusicaCommand(client: WASocket, botInfo: Bot, message: Message, group? : Group){
-    const commandsData = getCommands(botInfo)
+    const utilityCommands = commandsUtility(botInfo)
     const acr_host = process.env.acr_host
     const acr_access_key = process.env.acr_access_key
     const acr_access_secret = process.env.acr_access_secret
 
-    if (!acr_host || !acr_access_key || !acr_access_secret) throw new Error(commandsData.utility.qualmusica.msgs.error_key)
+    if (!acr_host || !acr_access_key || !acr_access_secret) throw new Error(utilityCommands.qualmusica.msgs.error_key)
 
     let messageType = message.isQuoted ? message.quotedMessage?.type : message.type
 
@@ -358,18 +357,18 @@ export async function qualmusicaCommand(client: WASocket, botInfo: Bot, message:
 
     let messageData = message.isQuoted ? message.quotedMessage?.wa_message : message.wa_message 
     
-    if (!messageData) throw new Error(commandsData.utility.qualmusica.msgs.error_message)
+    if (!messageData) throw new Error(utilityCommands.qualmusica.msgs.error_message)
 
     let messageMediaBuffer = await downloadMediaMessage(messageData, "buffer", {})
 
-    await Whatsapp.replyText(client, message.chat_id, commandsData.utility.qualmusica.msgs.wait, message.wa_message, {expiration: message.expiration})
+    await Whatsapp.replyText(client, message.chat_id, utilityCommands.qualmusica.msgs.wait, message.wa_message, {expiration: message.expiration})
     const musicResult = await audioLibrary.musicRecognition(messageMediaBuffer, {acr_host, acr_access_key, acr_access_secret})
-    const replyText = buildText(commandsData.utility.qualmusica.msgs.reply, musicResult.title, musicResult.producer, musicResult.duration, musicResult.release_date, musicResult.album, musicResult.artists)
+    const replyText = buildText(utilityCommands.qualmusica.msgs.reply, musicResult.title, musicResult.producer, musicResult.duration, musicResult.release_date, musicResult.album, musicResult.artists)
     await Whatsapp.replyText(client, message.chat_id, replyText, message.wa_message, {expiration: message.expiration})
 }
 
 export async function qualanimeCommand(client: WASocket, botInfo: Bot, message: Message, group? : Group){
-    const commandsData = getCommands(botInfo)
+    const utilityCommands = commandsUtility(botInfo)
 
     const messageData = {
         type: (message.isQuoted)? message.quotedMessage?.type : message.type,
@@ -377,15 +376,15 @@ export async function qualanimeCommand(client: WASocket, botInfo: Bot, message: 
     }
 
     if (messageData.type != "imageMessage") throw new Error(messageErrorCommandUsage(botInfo, message))
-    if (!messageData.message) throw new Error(commandsData.utility.qualanime.msgs.error_message)
+    if (!messageData.message) throw new Error(utilityCommands.qualanime.msgs.error_message)
 
-    await Whatsapp.replyText(client, message.chat_id, commandsData.utility.qualanime.msgs.wait, message.wa_message, {expiration: message.expiration})
+    await Whatsapp.replyText(client, message.chat_id, utilityCommands.qualanime.msgs.wait, message.wa_message, {expiration: message.expiration})
     const imageBuffer = await downloadMediaMessage(messageData.message, "buffer", {})
     const animeInfo = await imageLibrary.animeRecognition(imageBuffer)
 
-    if (animeInfo.similarity < 87) throw new Error(commandsData.utility.qualanime.msgs.error_similarity)
+    if (animeInfo.similarity < 87) throw new Error(utilityCommands.qualanime.msgs.error_similarity)
 
-    const replyText = buildText(commandsData.utility.qualanime.msgs.reply, animeInfo.title, animeInfo.episode || "---", animeInfo.initial_time, animeInfo.final_time, animeInfo.similarity, animeInfo.preview_url)
+    const replyText = buildText(utilityCommands.qualanime.msgs.reply, animeInfo.title, animeInfo.episode || "---", animeInfo.initial_time, animeInfo.final_time, animeInfo.similarity, animeInfo.preview_url)
     await Whatsapp.replyFileFromUrl(client, message.chat_id, 'videoMessage', animeInfo.preview_url, replyText, message.wa_message, {expiration: message.expiration, mimetype: 'video/mp4'})
 }
