+++ src/interfaces/command.interface.ts
@@ -3,7 +3,7 @@ import { Bot } from "./bot.interface.js"
 import { Message } from "./message.interface.js"
 import { Group } from "./group.interface.js"
 
-export type CategoryCommand = "info" | "utility" | "download" | "fun" | "group" | "sticker" | "admin"
+export type CategoryCommand = "info" | "utility" | "download" | "misc" | "group" | "sticker" | "admin"
 type CommandFunction = (client: WASocket, botInfo: Bot, message: Message, group?: Group) => Promise<void>
 
 export type Commands = {
